% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation.R
\name{bias_mitigation}
\alias{bias_mitigation}
\title{Bias mitigation}
\usage{
bias_mitigation(
  method,
  data_input,
  target_variable,
  protected_variable,
  probs = NULL,
  lambda = 1,
  cutoff = 0.5
)
}
\arguments{
\item{method}{character, Name of the mitigation method
1. "disparate-impact-remover"
2. "reweighing"
3. "resampling"}

\item{data_input}{\code{data.frame}, data to be transformed}

\item{target_variable}{character, target variable}

\item{protected_variable}{character, data column name which contains sensitive information such as gender, race etc...}

\item{lambda}{numeric, amount of repair desired. Value from 0 to 1, where 0 will return almost unchanged dataset and 1 fully repaired dataset}
}
\value{
Data/indices/weights after the mitigation (list)
}
\description{
Bias mitigation is a wrapper with several options to make a dataset more balanced.
}
\examples{

set.seed(1)
# custom data frame with kind, score and target column names
custom_data <- data.frame(
  kind = as.factor(c(rep("second", 500), rep("first", 500))),
  score = c(rnorm(500, 400, 40), rnorm(500, 600, 100))
  target = sample(c(0,1), replace=TRUE, size=1000)
)

mitigation_result <- bias_mitigation(
  method = disparate-impact-remover,
  data_input = custom_data,
  target_variable = custom_data$target,
  protected_variable = custom_data$kind,
  lambda = 0.8
)
}
