% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation.R
\name{disp_removing_data}
\alias{disp_removing_data}
\title{Disp Removing Data}
\usage{
disp_removing_data(data_input, target_variable, protected_variable, lambda)
}
\description{
This function mitigates bias with disparate impact remover pre-processing method (Feldman et al. (2015))
Filters out sensitive correlations in a dataset using 'disparate_impact_remover' function from fairmodels package.
}
