% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation.R
\name{resampling_data}
\alias{resampling_data}
\title{Resampling Data}
\usage{
resampling_data(
  method,
  ml_model,
  data_input,
  target_variable,
  protected_variablee,
  cutoff = 0.5
)
}
\description{
Resample the input data using fairmodels module function.
}
