% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_mitigation.R
\name{reweighing_model_weights}
\alias{reweighing_model_weights}
\title{Reweighing Model Weights}
\usage{
reweighing_model_weights(data_input, target_variable, protected_variable)
}
\description{
Function returns weights for model training. The purpose of this weights is to mitigate bias in statistical parity.
Obtain weights for model training using 'reweight' function from fairmodels package.
}
